footer fix
WIth the addition of links to the footer, we now have a few uneven columns on mobile. I added some code to fix this, however, most pages were referencing "style.css?v=1.2" instead of the "style.css" I added the code to. The issue should now be resolved, keeping 2 columns on the same row, regardless of height.